requirements.txt
Flask==2.3.3
opencv-python-headless==4.8.1.78
mysql-connector-python==8.1.0
Pillow==10.0.0
numpy==1.24.3
Werkzeug==2.3.7